+++ Letterbook.Api/Mappers/Converters/IdUuid7Converter.cs
@@ -1,18 +1,8 @@
 using AutoMapper;
-using Letterbook.Api.Dto;
 using Letterbook.Core.Models;
 using Medo;
-using Profile = Letterbook.Core.Models.Profile;
 
-namespace Letterbook.Api.Mappers;
-
-internal class Guid7Converter : IMemberValueResolver<Mention, MentionDto, Uuid7, Uuid7>
-{
-	public Uuid7 Resolve(Mention source, MentionDto destination, Uuid7 sourceMember, Uuid7 destMember, ResolutionContext context)
-	{
-		return source.GetId();
-	}
-}
+namespace Letterbook.Api.Mappers.Converters;
 
 internal class IdUuid7Converter :
 	ITypeConverter<IFederated, Uuid7>,
