+++ Letterbook.Api/Mappers/PostMappings.cs
@@ -1,8 +1,8 @@
 using AutoMapper;
 using Letterbook.Api.Dto;
+using Letterbook.Api.Mappers.Converters;
 using Letterbook.Core;
 using Letterbook.Core.Models;
-using Medo;
 using Microsoft.Extensions.Options;
 using Models = Letterbook.Core.Models;
 
@@ -75,41 +75,5 @@ public class PublicKeyPemConverter : IValueResolver<SigningKey, PublicKeyDto, st
 			SigningKey.KeyFamily.EcDsa => source.GetEcDsa().ExportSubjectPublicKeyInfoPem(),
 			_ => null!,
 		};
-	}
-}
-
-public class DateTimeOffsetMapper : IValueConverter<DateTimeOffset?, DateTimeOffset>,
-	IValueConverter<DateTimeOffset?, DateTimeOffset?>,
-	ITypeConverter<DateTimeOffset, DateTimeOffset>,
-	ITypeConverter<DateTimeOffset?, DateTimeOffset?>,
-	ITypeConverter<DateTimeOffset, DateTimeOffset?>,
-	ITypeConverter<DateTimeOffset?, DateTimeOffset>
-{
-	private DateTimeOffset ToUtc(DateTimeOffset? source) => source?.ToUniversalTime() ?? DateTimeOffset.UtcNow;
-	private DateTimeOffset? ToNullable(DateTimeOffset? source) => source == null ? null : ToUtc(source);
-
-	public DateTimeOffset Convert(DateTimeOffset? sourceMember, ResolutionContext context) => ToUtc(sourceMember);
-
-	DateTimeOffset? IValueConverter<DateTimeOffset?, DateTimeOffset?>.Convert(DateTimeOffset? sourceMember, ResolutionContext context)
-		=> ToNullable(sourceMember);
-
-	public DateTimeOffset Convert(DateTimeOffset source, DateTimeOffset destination, ResolutionContext context)
-		=> ToUtc(source);
-
-	public DateTimeOffset? Convert(DateTimeOffset? source, DateTimeOffset? destination, ResolutionContext context)
-		=> ToNullable(source);
-
-	public DateTimeOffset? Convert(DateTimeOffset source, DateTimeOffset? destination, ResolutionContext context)
-		=> ToNullable(source);
-
-	public DateTimeOffset Convert(DateTimeOffset? source, DateTimeOffset destination, ResolutionContext context)
-		=> ToUtc(source);
-}
-
-public class BaseTypeMappings : AutoMapper.Profile
-{
-	public BaseTypeMappings()
-	{
-
 	}
 }
